requirements.txt
addict==2.2.0
alembic==1.4.2
apache-airflow==1.10.11
apipkg==1.5
apispec==1.3.3
appdirs==1.4.4
appnope==0.1.0
argcomplete==1.12.0
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
bagit==1.7.0
blessed==1.17.9
CacheControl==0.11.7
cached-property==1.5.1
cattrs==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
colorama==0.4.3
coloredlogs==14.0
colorlog==4.0.2
configparser==3.5.3
connexion==2.7.0
contextlib2==0.6.0.post1
coverage==5.2.1
croniter==0.3.34
cwl-airflow==1.2.2
cwl-utils==0.7
cwlref-runner==1.0
cwltest==2.0.20200220223835
cwltool==3.0.20200807132242
dask==2.23.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.1
dnspython==2.0.0
docker==2.5.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.16
email-validator==1.1.1
enlighten==1.6.1
execnet==1.7.1
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
graphviz==0.14.1
gunicorn==19.10.0
humanfriendly==8.2
idna==2.10
inflection==0.5.0
iniconfig==1.0.1
ipython==7.17.0
ipython-genutils==0.2.0
iso8601==0.1.12
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.10.3
json-merge-patch==0.2
jsonmerge==1.7.0
jsonschema==3.2.0
junit-xml==1.9
lazy-object-proxy==1.5.1
lockfile==0.12.2
lxml==4.5.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
mistune==0.8.4
mock==4.0.2
more-itertools==8.4.0
mypy-extensions==0.4.3
natsort==7.0.1
networkx==2.4
numpy==1.19.1
openapi-spec-validator==0.2.9
packaging==20.4
pandas==1.1.0
parso==0.7.1
path==15.0.0
path.py==12.5.0
pathlib2==2.3.2
pathtools==0.1.2
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prison==0.1.3
prompt-toolkit==3.0.6
prov==1.5.1
psutil==5.7.2
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
pytest-fixture-config==1.7.0
pytest-shutil==1.7.0
pytest-testmon==1.0.3
pytest-virtualenv==1.7.0
pytest-watch==4.2.0
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
rdflib==4.2.2
rdflib-jsonld==0.5.0
requests==2.24.0
ruamel.yaml==0.16.5
schema-salad==7.0.20200811075006
setproctitle==1.1.10
sh==1.13.1
shellescape==3.4.1
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
swagger-ui-bundle==0.0.8
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.3
thrift==0.13.0
toil==4.2.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.10
virtualenv==20.0.31
watchdog==0.10.3
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==0.16.1
WTForms==2.3.3
zope.deprecation==4.4.0
